Restructured `note-item.html` to allow links within the note content.
The note surface is no longer clickable and there is now an icon link
to open the note 'detail' view.

I experimented with keeping the previous structure and replacing the outer
anchor with a clickable `div` or `button` at runtime via javascript.
It worked but caused issues with accessibility which made it clear that
it was the wrong approach.