Removed dependency `@tailwindcss/typography`, replaced with internal typography.css.
This is almost entirely a clone of the tailwind typography styles, but with its
structure tweaked to make it easier to use.

All styles are defined with in terms of `.prose-*` (global classname)
and `.prose *` (.prose wrapper class), so any style can be
used in a 'standalone' way by just using the global classname on an element.

Selectors used are also much simpler than the tailwind version.