Refactor: `AppPost`, `ChangelogPost` now extend `Post` instead of `BasePost`.
i.e. multi-table inheritance rather than ABC inheritance

As a result, `get_feed` is now a simple queryset instead of a list that needs
to be built by chaining in python, resulting in a much faster response time.

I thought that ABC inheritance was the simpler option for this but have been
thoroughly disabused of that notion.