Revert "Updated node modules"
This reverts commit 7229b28ecd4759d7cd2ad82acb0a452c6b036564.

Something about this change caused an issue in `globals.css`:

```
@utility hover-surface-* {
  ...
  &:hover:not(:disabled), &:focus-visible:not(:disabled) {
    background-color: color-mix(in srgb, var(--hover, var(--fg)) 10%, var(--bg));
  }                                                             ^^^^^
}
```

When doing a production build, the opacity modifier gets removed somehow?
This results in my subtly themed hover backgrounds to become very unsubtle.
Doesn't happen in dev build.
Guessing this is something in tailwind but might be some other part of the
build process. Needs investigation.