Use explicit user group ID to ensure mounted media directory is writeable by django user.
Requires equivalent (mapped, according to `/etc/subgid`) `gid` to have ownership
of media directory on the host.

i.e.
```sh
> cat /etc/subgid/
username:100000:65536
```

The `gid` 19283 inside the container will map to `gid` 119282 on the host.

(This is only required, I believe, because we are using rootless docker with
`USER` definitions in our images.)